feat: enhance navbar components with responsive styles and new mobile sidebar toggle
